requirements.txt
openai==1.52.2
pinecone==5.3.1
python-dotenv==1.0.1
